Move RemoveSubmissionView tests to their own file.
One of many.
